"lakefs setup" to set up ACL groups (#5473)
* [refactor] git mv pkg/auth/setup.go pkg/auth/setup

setup now needs to access pkg/auth and pkg/auth/acl.  Need a new package to
prevent import cycle.

* "lakefs setup" to set up ACL groups

Depending on auth.ui_config.rbac "simplified" or "external", create
appropriate initial groups.

* dummy commit to bump CI

* [bug] Use new group names to run Esti

Now tests ACL-based setup (which is the default).

* [CR] Remove simple constants, reformat, delete fixed comment
